+++ src/IO/ftiff.h
@@ -96,9 +96,9 @@ typedef struct FTIFF_headInfo {
     int exists;
     int nbVar;
     char ckptFile[FTI_BUFS];
-    uint64_t maxFs;
-    uint64_t fs;
-    uint64_t pfs;
+    int64_t maxFs;
+    int64_t fs;
+    int64_t pfs;
     int isDcp;
 } FTIFF_headInfo;
 
@@ -152,7 +152,7 @@ typedef struct FTIFF_MPITypeInfo {
 
  **/
 int FTI_FinalizeFtiff(void *fd);
-size_t FTI_DummyFilePos(void *ignore);
+int64_t FTI_DummyFilePos(void *ignore);
 int FTI_WriteFtiffData(FTIT_dataset* data, void *fd);
 void* FTI_InitFtiff(FTIT_configuration* FTI_Conf, FTIT_execution* FTI_Exec,
         FTIT_topology* FTI_Topo, FTIT_checkpoint* FTI_Ckpt,
@@ -206,7 +206,7 @@ void FTIFF_SetHashChunk(FTIFF_dbvar *dbvar, FTIT_keymap* FTI_Data);
 void FTIFF_PrintDataStructure(int rank, FTIT_execution* FTI_Exec);
 int FTI_ProcessDBVar(FTIT_execution *FTI_Exec, FTIT_configuration *FTI_Conf,
  FTIFF_dbvar *currentdbvar,  FTIT_dataset *data, unsigned char *hashchk,
- WriteFTIFFInfo_t *fd, uint64_t *dcpSize, unsigned char **dptr);
+ WriteFTIFFInfo_t *fd, int64_t *dcpSize, unsigned char **dptr);
 int FTIFF_RecoverVarInit(char* fn);
 int FTIFF_RecoverVarFinalize();
 #endif  // FTI_SRC_IO_FTIFF_H_
